requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.7
Pillow==5.1.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
